chore(model): refactor
