+++ .changeset/fifty-bananas-sleep.md
@@ -1,0 +1,5 @@
+---
+'@chikichat/model': patch
+---
+
+refactor
